requirements.txt
apache-airflow
apache-airflow-providers-oracle
cx_Oracle